// backend/app/src/gen/pb/api/v2/app/room/service/RoomService_rb.service.ts
// @generated by protoc-gen-restobuf-nestjs v1 with parameter "target=ts,import_extension=js"
// @generated from file api/v2/app/room/service.proto (package api.v2.app.room, syntax proto3)
/* eslint-disable */

import { Injectable } from "@nestjs/common";
import type { CreateRoomRequest, CreateRoomResponse, EnterRoomRequest, EnterRoomResponse, ExitRoomRequest, ExitRoomResponse, GetRoomRequest, GetRoomResponse, ResetRoomRequest, ResetRoomResponse } from "../service_pb.js";
import { Request, Response } from "express";

@Injectable() 
export abstract class RoomServiceService {
  abstract handleCreateRoom(input: CreateRoomRequest, req: Request, res: Response): Promise<CreateRoomResponse>;
  abstract handleGetRoom(input: GetRoomRequest, req: Request, res: Response): Promise<GetRoomResponse>;
  abstract handleResetRoom(input: ResetRoomRequest, req: Request, res: Response): Promise<ResetRoomResponse>;
  abstract handleEnterRoom(input: EnterRoomRequest, req: Request, res: Response): Promise<EnterRoomResponse>;
  abstract handleExitRoom(input: ExitRoomRequest, req: Request, res: Response): Promise<ExitRoomResponse>;
}